+++ lib/qonfig.rb
@@ -63,7 +63,7 @@ module Qonfig
   # @since 0.20.0
   define_validator(:module) { |value| value.is_a?(Module) }
   # @since 0.20.0
-  define_validator(:not_nil) { |value| !value.nil? }
+  define_validator(:not_nil) { |value| value == nil }
 
   # @since 0.26.0
   define_resolver(:file) do |file_path|