+++ spec/features/plugins/vault/context.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-shared_context "vault context" do
+shared_context 'vault context' do
   before { allow(Vault).to receive(:logical).and_return(logical_double) }
 
   before { allow(Vault).to receive(:kv).and_return(kv_double) }