+++ spec/features/validation_spec.rb
@@ -236,6 +236,7 @@ describe 'Validation' do
       begin
         config.clear! # NOTE: set telegraf_url to nil (to incorrect value)
       rescue Qonfig::ValidationError
+        # NOOP
       end
 
       expect(config.valid?).to eq(false)
@@ -300,6 +301,7 @@ describe 'Validation' do
       begin
         config.settings.adapter = 123
       rescue Qonfig::ValidationError
+        # NOOP
       end
       expect(config.valid?).to eq(false)
       expect(config.settings.adapter).to eq(123)
@@ -318,6 +320,7 @@ describe 'Validation' do
       begin
         config.settings.enabled = '123'
       rescue Qonfig::ValidationError
+        # NOOP
       end
       expect(config.valid?).to eq(false)
       expect(config.settings.enabled).to eq('123')
@@ -804,7 +807,7 @@ describe 'Validation' do
 
         # custom class-related validator
         define_validator(:queue_adapter) do |value|
-          value == :que || value == :sidekiq # rubocop:disable Style/MultipleComparison
+          value == :que || value == :sidekiq
         end
 
         setting :some_value, 123